+++ index.js
@@ -4,16 +4,16 @@
 
 // Uncomment lines 7-14 to run Example App
 
-import { AppRegistry } from 'react-native';
-import { name as appName } from './app.json';
-import App from './ExampleApp/App';
+// import { AppRegistry } from 'react-native';
+// import { name as appName } from './app.json';
+// import App from './ExampleApp/App';
 
-const debug = require('debug');
-debug.enable('rnMatrix:*');
+// const debug = require('debug');
+// debug.enable('rnm:*');
 
-console.disableYellowBox = true;
+// console.disableYellowBox = true;
 
-AppRegistry.registerComponent(appName, () => App);
+// AppRegistry.registerComponent(appName, () => App);
 
 //********************************************************************************
 // NPM Module